Respect the Minitest convention for test directory
Currently, when you use this gem with RSpec, you can have all
test-related code segregated to the `spec/` folder since it looks for
`spec/doctest_helper.rb`. This is not true if you use Minitest since it
doesn't look in the conventional `test/` directory that Minitest uses.

To make the gem work well with Minitest as well, this change allows the
helper file to live in the `test/` directory.