Allow SimpleCov to correctly collect coverage
Because the original implementation required `minitest/autorun` in front
of every test instance, Minitest eagerly ran each test as it was
defined, before the other tests were run. This causes modern versions of
SimpleCov to calculate coverage too early in the process.

This change explicitly runs Minitest after compiling every example. This
is more in-line with how normal test suites run so SimpleCov works like
normal as long as you set it up in the `doctest_helper.rb` file.